+++ changelog.md
@@ -5,6 +5,7 @@
 ### Patch
 
 - Updated dependencies.
+- Allow batched operations again, fixing [#142](https://github.com/jaydenseric/graphql-upload/issues/142).
 - Simplify tests by writing JSON as strings instead of using `JSON.stringify`.
 - Use async middleware for Express tests.
 - Added the Open Graph image design to the logo Sketch file.